Techcrunch
Too difficult to get this to work.  Seems that techcrunch checks for a cookie and if it isn't there forwards requests to guce.techcrunch.com which forwards on to guce.advertising.com which in turn forwards the request to a whole bunch of Yahoo properties before finally returning you to the article.  Just going to whitelist guce.advertising.com in nextdns.